moved commands utils to behaviour
